requirements.txt
pbr>=0.5.21,<1.0

PyYAML
python-jenkins
paramiko
lockfile
python-daemon
extras
statsd>=1.0.0,<3.0
voluptuous>=0.6,<0.7
apscheduler>=2.1.1,<3.0
sqlalchemy>=0.8.2,<0.9.0
pyzmq>=13.1.0,<14.0.0
python-novaclient
